+++ source/relevant_document.h
@@ -14,13 +14,13 @@ class ANT_relevant_document_passage;
 class ANT_relevant_document
 {
 public:
-	long topic;
+	long long topic;
 	long long docid;
-	long relevant_characters;
-	long document_length;
+	long long relevant_characters;
+	long long document_length;
 	ANT_relevant_document_passage *passage_list;
-	long number_of_relevant_passages;
-	long best_entry_point;
+	long long number_of_relevant_passages;
+	long long best_entry_point;
 
 public:
 	static int compare(const void *a, const void *b);
